# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLK Pro API",
    "version": "v1",
    "description": "Auto-generated from routes/api.php (internal API v1)."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "AuthController@logout",
        "operationId": "AuthController_logout_post",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\Api\\AuthController@logout",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "AuthController@me",
        "operationId": "AuthController_me_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\Api\\AuthController@me",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended"
        ]
      }
    },
...
    "/api/v1/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "ProfileController@edit",
        "operationId": "ProfileController_edit_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\ProfileController@edit",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete"
        ]
      },
      "patch": {
        "tags": [
          "profile"
        ],
        "summary": "ProfileController@update",
        "operationId": "ProfileController_update_patch",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\ProfileController@update",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete"
        ]
      },
      "delete": {
        "tags": [
          "profile"
        ],
        "summary": "ProfileController@destroy",
        "operationId": "ProfileController_destroy_delete",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\ProfileController@destroy",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete"
        ]
      }
    },
...
    "/api/v1/services/categories": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "ServiceController@categories",
        "operationId": "ServiceController_categories_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\ServiceController@categories",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:services"
        ]
      }
    },
    "/api/v1/services/options": {
      "get": {
        "tags": [
          "services"
        ],
        "summary": "ServiceController@options",
        "operationId": "ServiceController_options_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\ServiceController@options",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:services"
        ]
      }
    },
    "/api/v1/services/quick": {
      "post": {
        "tags": [
          "services"
        ],
        "summary": "ServiceController@storeQuick",
        "operationId": "ServiceController_storeQuick_post",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\ServiceController@storeQuick",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:services"
        ]
      }
    },
...
  }
}